Move writablestream into its own module
